drag-point: Move classes to separate namespace.
