+++ VisualPinball.Unity/VisualPinball.Unity.Editor/DragPoint/CatmullCurveHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd9d759cf209e2f4497eadd6a09399be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
